// src/types/recipe.ts
export interface Recipe {
  idMeal: string;
  strMeal: string;
  strDrinkAlternate: string | null;
  strCategory: string;
  strArea: string;
  strInstructions: string;
  strMealThumb: string;
  strTags: string | null;
  strYoutube: string;
  strSource: string;
  strImageSource: string | null;
  strCreativeCommonsConfirmed: string | null;
  dateModified: string | null;
  ingredients: Ingredient[];
}

export interface Ingredient {
  ingredient: string;
  measure: string;
}

export interface RecipeResponse {
  meals: Recipe[];
}